## Task
The client sends transaction reports in the form of text messages, which contain information about requests for information (RFI), chargebacks, and pending refunds (Chargebacks Pending). It is necessary to automate the process of extracting data from these emails and entering it into a table for easy analysis and tracking.

## Solution

To solve this problem, the Python programming language is used. Using the pandas library for working with data and regular expressions to extract the necessary information, a script was written that analyzes text messages, extracts data from the REQUESTS FOR INFORMATION REPORT section, and adds it to the appropriate table. This process allows incoming reports to be processed efficiently, providing a clear view of the data for further analysis and monitoring.